remove unused method of Auth interface
